add Marshaler and Unmarshaler interfaces
for custom Marshal and Unmarshal functionality. Right now, the only
use for this is the RawMessage type. The Marshal functionality is not
yet implemented...

Unmarshal functionality is implemented by passing along a *[]byte
through all the decode calls.  probably can be optimized by passing
around offsets instead of constructing a new byte array.