rework interface and implementation for listening for unilateral messages
Rather than having two send methods - Send and SendAndReceive - we
remove the latter, and add a new Receive method for listening to all
unilateral messages.

As the first usage, we also add a LogEvent type and a test for the
LogLevel which exerises listening for log messages.

The implementation has been altered to overload the reqCh to handle
multiple types of requests so that we can send messages, receive
requests to listen for unilateral messages and receive requests to stop
listening in a thread-safe manner.