fix rawmessage decoding of slices
68bda102 added some logic to ensure that we never write past the len of
a slice or array when decoding. unfortunately, it didn't check if the
dest was valid - which is important for handling the rawmessage case.

subtly, i think this logic was wrong in the normal case, because we need
to continue reading from the byte stream even if we can't set the value,
so that further values that are decoded start from the correct position.

instead, treat this case (where i > len(dest)) as if we had an invalid
dest, which just keeps reading, but skips setting.