[bser] add tests + bench for Marshaler+Unmarshaler
This adds handling for encoding data implementing the
Marshaler interface and adds some basic tests.

Since so far we had only tested the RawMessage implementation
of the Unmarshaler interface, an additional test case
was added to verify behaviour for a different implementation.

Benchmark cases were also added to help establish a
base for measuring performance.